fix(nlsq): prevent incomplete chunks in angle-stratified optimization
Resolves critical bug where stratification created chunks missing
certain phi angles, causing zero gradients and silent optimization
failures with per-angle scaling on large datasets (>1M points).

Root cause: Algorithm continued creating chunks after some angles
were exhausted, violating the requirement that ALL chunks must
contain ALL phi angles for gradient computation.

Solution: Calculate max safe chunks upfront based on minimum angle
size, ensuring all chunks are complete before any angle exhausts.
Adds epsilon values to prevent division by zero in sequential angle
optimization and fixes dtype consistency in JAX operations.

Changes:
- stratified_chunking.py: Replace while loop with bounded iteration
- sequential_angle.py: Add epsilon (1e-10) to division operations
- jax_backend.py: Ensure dtype consistency in array operations
- pytest.ini: Add filterwarnings, remove unsupported options
- pyproject.toml: Register custom pytest markers

Impact: Stratified least_squares now succeeds instead of falling
back to curve_fit_large for datasets with 23M+ points and 23 angles.

+++ homodyne/core/jax_backend.py
@@ -864,7 +864,8 @@ def apply_diagonal_correction(c2_mat: jnp.ndarray) -> jnp.ndarray:
 
     # Compute diagonal values as average of adjacent off-diagonal elements
     # This implementation matches xpcs_loader.py:924-953 but uses pure JAX ops
-    diag_val = jnp.zeros(size)
+    # Use same dtype as input matrix to avoid casting warnings
+    diag_val = jnp.zeros(size, dtype=c2_mat.dtype)
 
     # Add left neighbors: diag_val[:-1] += side_band
     diag_val = diag_val.at[:-1].add(side_band)
@@ -873,7 +874,7 @@ def apply_diagonal_correction(c2_mat: jnp.ndarray) -> jnp.ndarray:
     diag_val = diag_val.at[1:].add(side_band)
 
     # Normalize by number of neighbors (1 for edges, 2 for middle)
-    norm = jnp.ones(size)
+    norm = jnp.ones(size, dtype=c2_mat.dtype)
     norm = norm.at[1:-1].set(2.0)  # Middle elements have 2 neighbors
 
     diag_val = diag_val / norm
